# packages/shared-schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisRunResponse": {
        "additionalProperties": false,
        "properties": {
          "algorithm_version": {
            "title": "Algorithm Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AnalysisRunKind"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "random_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Random Seed"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisRunStatus"
          },
          "surface_dataset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface Dataset Id"
          }
        },
        "required": [
          "id",
          "project_id",
          "surface_dataset_id",
          "kind",
          "status",
          "algorithm_version",
          "parameters",
          "random_seed",
          "metrics",
          "error",
          "started_at",
          "finished_at",
          "created_at"
        ],
        "title": "AnalysisRunResponse",
        "type": "object"
      },
      "AnalysisRunStatus": {
        "enum": [
          "pending",
          "running",
          "completed",
          "failed"
        ],
        "title": "AnalysisRunStatus",
        "type": "string"
      },
...
      "CandidateListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateSiteResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "CandidateListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ViewshedRunRequest": {
        "additionalProperties": false,
        "description": "Parameters for a batch viewshed computation over a candidate run.",
        "properties": {
          "candidate_site_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Compute only these candidates; defaults to every accepted site.",
            "title": "Candidate Site Ids"
          },
          "max_distance_m": {
            "default": 10000.0,
            "maximum": 50000.0,
            "minimum": 100.0,
            "title": "Max Distance M",
            "type": "number"
          },
          "observer_height_m": {
            "default": 10.0,
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Observer Height M",
            "type": "number"
          },
          "refraction_coefficient": {
            "default": 0.13,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Refraction Coefficient",
            "type": "number"
          },
          "target_height_m": {
            "default": 0.0,
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Target Height M",
            "type": "number"
          },
          "use_earth_curvature": {
            "default": true,
            "title": "Use Earth Curvature",
            "type": "boolean"
          }
        },
        "title": "ViewshedRunRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Geospatial planning of wildfire surveillance tower networks.",
    "title": "Sentinel Planner API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analysis-runs/{candidates_run_id}/viewsheds": {
      "post": {
        "description": "Queue viewsheds for every accepted candidate of a run (or a subset).\n\nReturns immediately with status ``pending`` or ``completed`` (if every\nrequested viewshed was already cached) \u2014 nothing is computed inside this\nrequest. A worker process picks up pending work; see\n``app.workers.viewshed_worker``.",
        "operationId": "enqueue_viewsheds_api_v1_analysis_runs__candidates_run_id__viewsheds_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidates_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidates Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewshedRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue viewshed computation for a candidate run's sites",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/analysis-runs/{run_id}": {
      "get": {
        "operationId": "get_analysis_run_api_v1_analysis_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get an analysis run",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/analysis-runs/{run_id}/candidates": {
      "get": {
        "operationId": "list_candidates_api_v1_analysis_runs__run_id__candidates_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "allowed_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Allowed Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the candidate sites of a run",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/analysis-runs/{run_id}/optimization-solutions": {
      "get": {
        "operationId": "list_optimization_solutions_api_v1_analysis_runs__run_id__optimization_solutions_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationSolutionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the optimization solutions of a run",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/analysis-runs/{run_id}/viewsheds": {
      "get": {
        "operationId": "list_viewsheds_api_v1_analysis_runs__run_id__viewsheds_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewshedListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the viewsheds of a batch run, with progress",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/analysis-runs/{viewshed_run_id}/optimize": {
      "post": {
        "description": "Greedily choose candidates from a completed (or partial) viewshed run.\n\nRuns synchronously: this operates on masks Phase 3 already computed \u2014\narray operations, not ray casting \u2014 so it finishes in seconds even for\nhundreds of candidates.",
        "operationId": "optimize_coverage_api_v1_analysis_runs__viewshed_run_id__optimize_post",
        "parameters": [
          {
            "in": "path",
            "name": "viewshed_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Viewshed Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationSolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select the Sentinel positions that maximize covered surface",
        "tags": [
          "analysis-runs"
        ]
      }
    },
...
    "/api/v1/optimization-solutions/{solution_id}": {
      "get": {
        "operationId": "get_optimization_solution_api_v1_optimization_solutions__solution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "solution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Solution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationSolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a single optimization solution",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/optimization-solutions/{solution_id}/export.csv": {
      "get": {
        "operationId": "export_solution_csv_api_v1_optimization_solutions__solution_id__export_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "solution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Solution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/csv": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the selected Sentinel positions as CSV",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/optimization-solutions/{solution_id}/export.geojson": {
      "get": {
        "operationId": "export_solution_geojson_api_v1_optimization_solutions__solution_id__export_geojson_get",
        "parameters": [
          {
            "in": "path",
            "name": "solution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Solution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/geo+json": {},
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the selected Sentinel positions as GeoJSON",
        "tags": [
          "analysis-runs"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/analysis-runs": {
      "get": {
        "operationId": "list_analysis_runs_api_v1_projects__project_id__analysis_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AnalysisRunKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List a project's analysis runs",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/candidates": {
      "post": {
        "description": "Grid the study area, filter by terrain, thin by separation, and persist.\n\nRuns against the project's most recent processed DEM. Fails with a clear\nmessage if no DEM has been ingested yet.",
        "operationId": "generate_candidates_api_v1_projects__project_id__candidates_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate candidate Sentinel positions",
        "tags": [
          "analysis-runs"
        ]
      }
    },
...
    "/api/v1/viewsheds/{viewshed_id}": {
      "get": {
        "operationId": "get_viewshed_api_v1_viewsheds__viewshed_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "viewshed_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Viewshed Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewshedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a single viewshed",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/viewsheds/{viewshed_id}/mask.tif": {
      "get": {
        "operationId": "download_viewshed_mask_api_v1_viewsheds__viewshed_id__mask_tif_get",
        "parameters": [
          {
            "in": "path",
            "name": "viewshed_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Viewshed Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/tiff": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download the viewshed mask GeoTIFF",
        "tags": [
          "analysis-runs"
        ]
      }
    },
    "/api/v1/viewsheds/{viewshed_id}/preview.png": {
      "get": {
        "operationId": "download_viewshed_preview_api_v1_viewsheds__viewshed_id__preview_png_get",
        "parameters": [
          {
            "in": "path",
            "name": "viewshed_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Viewshed Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download a map-overlay preview of the viewshed",
        "tags": [
          "analysis-runs"
        ]
      }
    }
  }
}